requirements.txt
numpy==1.18.1
pandas==1.0.1
protobuf==3.11.3
scipy==1.4.1
tensorboard==2.1.0
tensorflow-cpu==2.1.0

